requirements.txt
torch: 2.7.1+cpu
torchvision: 0.22.1+cpu
streamlit: 1.32.0
Pillow: 10.3.0
numpy: 1.26.4